requirements.txt
Django
djangorestframework
django-cors-headers
pymongo
djongo
bcrypt